Move uigo to ui
